Moved fixed models to model package and fixed usage in HelloController
